localization support

+++ src/main/java/dev/ambershadow/cogfly/elements/settings/ResetToDefaultElement.java
@@ -13,12 +13,15 @@ public class ResetToDefaultElement extends SettingsElement {
     public ResetToDefaultElement(SettingsDialog parent) {
         JButton button = new JButton("Reset To Default");
         button.addActionListener(_ -> {
-            try {
-                Files.delete(Cogfly.dataJson);
-            } catch (IOException e) {
-                throw new RuntimeException(e);
+            int a = JOptionPane.showConfirmDialog(parent, "This will reset ALL settings and persistent values, are you sure you want to do this?", "Confirm Reset", JOptionPane.YES_NO_OPTION, JOptionPane.QUESTION_MESSAGE);
+            if (a == JOptionPane.YES_OPTION) {
+                try {
+                    Files.delete(Cogfly.dataJson);
+                } catch (IOException e) {
+                    throw new RuntimeException(e);
+                }
+                parent.setAndClose(Settings.load(Cogfly.dataJson));
             }
-            parent.setAndClose(Settings.load(Cogfly.dataJson));
         });
         button.setToolTipText("Reset all settings to their default.");
         button.setPreferredSize(new Dimension(625, button.getPreferredSize().height));
